pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "media-archiver"
version = "0.1.0"
description = "Deterministic, offline-only photo and video archive organizer"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Private" }
authors = [{ name = "Project Owner" }]

dependencies = [
    "PyYAML>=6.0",
    "Pillow>=10.0", # EXIF extraction
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[project.scripts]
media-archiver = "media_archiver.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.coverage.run]
branch = true
source = ["media_archiver"]

[tool.coverage.report]
show_missing = true
